+++ flow-typed/core-components.js
@@ -29,11 +29,27 @@ declare module '@atlaskit/dynamic-table' {
   declare module.exports: any;
 }
 
+declare module '@atlaskit/flag' {
+  declare module.exports: any;
+}
+
 declare module '@atlaskit/inline-edit' {
   declare module.exports: any;
 }
 
+declare module '@atlaskit/lozenge' {
+  declare module.exports: any;
+}
+
+declare module '@atlaskit/page' {
+  declare module.exports: any;
+}
+
 declare module '@atlaskit/pagination' {
+  declare module.exports: any;
+}
+
+declare module '@atlaskit/section-message' {
   declare module.exports: any;
 }