+++ packages/core/global-navigation/src/components/GlobalNavigation/index.js
@@ -418,7 +418,6 @@ export default class GlobalNavigation extends Component<
       product,
       cloudId,
       /* eslint-disable camelcase */
-      experimental_enableSplitJira,
       experimental_enableExpandLink,
       /* eslint-enable camelcase */
     } = this.props;
@@ -428,7 +427,6 @@ export default class GlobalNavigation extends Component<
         product={product}
         triggerXFlow={this.triggerXFlow}
         /* eslint-disable camelcase */
-        enableSplitJira={experimental_enableSplitJira}
         enableExpandLink={experimental_enableExpandLink}
         /* eslint-enable camelcase */
       />