+++ packages/editor/editor-core/src/__tests__/visual-regression/full-width-mode/full-width-mode.ts
@@ -1,3 +1,4 @@
+import { MINIMUM_THRESHOLD } from '@atlaskit/visual-regression/helper';
 import { initFullPageEditorWithAdf, Device, snapshot } from '../_utils';
 import adf from './__fixtures__/mixed-content.adf.json';
 import adfWithMedia from './__fixtures__/content-with-media.adf.json';
@@ -26,7 +27,7 @@ widths.forEach(width => {
 
     it('should display content in full-width mode', async () => {
       await scrollToTop(page);
-      await snapshot(page, 0.02);
+      await snapshot(page, MINIMUM_THRESHOLD);
     });
   });
 
@@ -43,7 +44,7 @@ widths.forEach(width => {
         { appearance: 'full-width' },
       );
       await waitForLoadedImageElements(page);
-      await snapshot(page, 0.02);
+      await snapshot(page, MINIMUM_THRESHOLD);
     });
   });
 });
@@ -59,6 +60,6 @@ describe('Full-width mode breakout', () => {
       { width: 2000, height: 800 },
       { appearance: 'full-width' },
     );
-    await snapshot(page, 0.02);
+    await snapshot(page, MINIMUM_THRESHOLD);
   });
 });