+++ packages/search/global-search/src/components/GlobalQuickSearchWrapper.tsx
@@ -6,7 +6,11 @@ import JiraQuickSearchContainer from './jira/JiraQuickSearchContainer';
 import configureSearchClients, { Config } from '../api/configureSearchClients';
 import MessagesIntlProvider from './MessagesIntlProvider';
 import { GlobalSearchPreFetchContext } from './PrefetchedResultsProvider';
-import { QuickSearchContext } from '../api/types';
+import {
+  QuickSearchContext,
+  ConfluenceModelContext,
+  JiraModelContext,
+} from '../api/types';
 
 const DEFAULT_NOOP_LOGGER: Logger = {
   safeInfo() {},
@@ -148,6 +152,16 @@ export interface Props {
    * This is used for Confluence only.
    */
   fasterSearchFFEnabled?: boolean;
+
+  /**
+   * Determine whether to enable urs for bootstrapping people search.
+   */
+  useUrsForBootstrapping?: boolean;
+
+  /**
+   * Additional context paramters used to evaluate search models
+   */
+  modelContext?: ConfluenceModelContext | JiraModelContext;
 }
 
 /**
@@ -252,6 +266,8 @@ export default class GlobalQuickSearchWrapper extends React.Component<Props> {
               inputControls,
               appPermission,
               fasterSearchFFEnabled,
+              useUrsForBootstrapping,
+              modelContext,
             } = this.props;
 
             return (
@@ -268,6 +284,8 @@ export default class GlobalQuickSearchWrapper extends React.Component<Props> {
                 inputControls={inputControls}
                 appPermission={appPermission}
                 fasterSearchFFEnabled={fasterSearchFFEnabled}
+                useUrsForBootstrapping={useUrsForBootstrapping}
+                modelContext={modelContext}
               />
             );
           }}