+++ packages/core/icon/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/icon
 
+## 17.1.1
+- Updated dependencies [6dd86f5b07](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/6dd86f5b07):
+  - @atlaskit/field-text@9.0.1
+  - @atlaskit/theme@9.0.2
+  - @atlaskit/section-message@4.0.0
+
 ## 17.1.0
 - [minor] [8f4f5d914a](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/8f4f5d914a):